+++ pilot/pkg/networking/core/v1alpha3/cluster.go
@@ -183,7 +183,9 @@ func buildClusterHosts(env *model.Environment, proxyNetworkView map[string]bool,
 
 	hosts := make([]*core.Address, 0)
 	for _, instance := range instances {
-		// Only send endpoints from the networks in the network view.
+		// Only send endpoints from the networks in the network view requested by the proxy.
+		// The default network view assigned to the Proxy is the UnnamedNetwork (""), which matches
+		// the default network assigned to endpoints that don't have an explicit network
 		if _, found := proxyNetworkView[instance.Endpoint.Network]; !found {
 			// Endpoint's network doesn't match the set of networks that the proxy wants to see.
 			continue
